requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
python-multipart>=0.0.9
aiohttp>=3.9.0
aiofiles>=23.0.0
pydantic>=2.6.0
python-dotenv>=1.0.0
requests>=2.31.0
